+++ tests/test_api_result.py
@@ -1,8 +1,10 @@
 import contextlib
-import functools
+import json
 import logging
+import shutil
 import textwrap
-from typing import List, Union, Sequence, Tuple
+from pathlib import Path
+from typing import List, Union, Sequence
 
 import mock
 import numpy as np
@@ -14,9 +16,25 @@ from shapely.geometry import GeometryCollection, Point, Polygon, box, mapping
 
 import openeo
 import openeo.processes
-from openeo_driver.testing import TEST_USER, ApiResponse, DictSubSet, IgnoreOrder
-from openeo_driver.util.geometry import as_geojson_feature, as_geojson_feature_collection
-from openeogeotrellis.testing import random_name
+from openeo_driver.jobregistry import JOB_STATUS
+from openeo_driver.testing import (
+    TEST_USER,
+    ApiResponse,
+    DictSubSet,
+    IgnoreOrder,
+    load_json,
+    ApiTester,
+    RegexMatcher,
+    ListSubSet,
+    UrllibMocker,
+)
+from openeo_driver.util.geometry import (
+    as_geojson_feature,
+    as_geojson_feature_collection,
+)
+from openeogeotrellis.backend import JOB_METADATA_FILENAME
+from openeogeotrellis.job_registry import ZkJobRegistry
+from openeogeotrellis.testing import random_name, KazooClientMock
 from openeogeotrellis.utils import (
     UtcNowClock,
     drop_empty_from_aggregate_polygon_result,
@@ -334,18 +352,20 @@ def test_udp_udf_reduce_temporal(api100, user_defined_process_registry, udf_code
 
 
 @pytest.mark.parametrize("set_offset", [False, True])
-@pytest.mark.parametrize("udf_code", [
-    """
+@pytest.mark.parametrize(
+    "udf_code",
+    [
+        """
         from openeo_udf.api.datacube import DataCube  # Old style openeo_udf API
         def apply_datacube(cube: DataCube, context: dict) -> DataCube:
             offset = context.get("offset", 34)
-            return DataCube(cube.get_array().max("t") + offset) 
+            return DataCube(cube.get_array().max("t") + offset)
     """,
-    """
+        """
         from openeo.udf import XarrayDataCube
         def apply_datacube(cube: XarrayDataCube, context: dict) -> XarrayDataCube:
             offset = context.get("offset", 34)
-            return XarrayDataCube(cube.get_array().max("t") + offset) 
+            return XarrayDataCube(cube.get_array().max("t") + offset)
     """,
 ])
 def test_udp_udf_reduce_temporal_with_parameter(api100, user_defined_process_registry, set_offset, udf_code):
@@ -414,24 +434,26 @@ def test_udp_udf_reduce_temporal_with_parameter(api100, user_defined_process_reg
 
 
 @pytest.mark.parametrize("set_parameters", [False, True])
-@pytest.mark.parametrize("udf_code", [
-    """
+@pytest.mark.parametrize(
+    "udf_code",
+    [
+        """
         from openeo_udf.api.datacube import DataCube  # Old style openeo_udf API
         def apply_datacube(cube: DataCube, context: dict) -> DataCube:
             l_scale = context.get("l_scale", 100)
             d_scale = context.get("d_scale", 1)
             array = cube.get_array()
-            res = l_scale * array.sel(bands="Longitude") + d_scale * array.sel(bands="Day") 
-            return DataCube(res) 
+            res = l_scale * array.sel(bands="Longitude") + d_scale * array.sel(bands="Day")
+            return DataCube(res)
     """,
-    """
+        """
         from openeo.udf import XarrayDataCube
         def apply_datacube(cube: XarrayDataCube, context: dict) -> XarrayDataCube:
             l_scale = context.get("l_scale", 100)
             d_scale = context.get("d_scale", 1)
             array = cube.get_array()
-            res = l_scale * array.sel(bands="Longitude") + d_scale * array.sel(bands="Day") 
-            return XarrayDataCube(res) 
+            res = l_scale * array.sel(bands="Longitude") + d_scale * array.sel(bands="Day")
+            return XarrayDataCube(res)
     """,
 ])
 def test_udp_udf_reduce_bands_with_parameter(api100, user_defined_process_registry, set_parameters, udf_code):
@@ -556,18 +578,20 @@ def test_apply_square_pixels(api100):
     assert_equal(data, expected)
 
 
-@pytest.mark.parametrize("udf_code", [
-    """
+@pytest.mark.parametrize(
+    "udf_code",
+    [
+        """
         from openeo_udf.api.datacube import DataCube  # Old style openeo_udf API
         def apply_datacube(cube: DataCube, context: dict) -> DataCube:
             array = cube.get_array()
-            return DataCube(array * array)  
+            return DataCube(array * array)
     """,
-    """
+        """
         from openeo.udf import XarrayDataCube
         def apply_datacube(cube: XarrayDataCube, context: dict) -> XarrayDataCube:
             array = cube.get_array()
-            return XarrayDataCube(array * array)  
+            return XarrayDataCube(array * array)
     """,
 ])
 def test_apply_udf_square_pixels(api100, udf_code):
@@ -618,18 +642,20 @@ def test_apply_udf_square_pixels(api100, udf_code):
 
 
 @pytest.mark.parametrize("set_parameters", [False, True])
-@pytest.mark.parametrize("udf_code", [
-    """
+@pytest.mark.parametrize(
+    "udf_code",
+    [
+        """
         from openeo_udf.api.datacube import DataCube  # Old style openeo_udf API
         def apply_datacube(cube: DataCube, context: dict) -> DataCube:
             offset = context.get("offset", 100)
-            return DataCube(cube.get_array() + offset) 
+            return DataCube(cube.get_array() + offset)
     """,
-    """
+        """
         from openeo.udf import XarrayDataCube
         def apply_datacube(cube: XarrayDataCube, context: dict) -> XarrayDataCube:
             offset = context.get("offset", 100)
-            return XarrayDataCube(cube.get_array() + offset) 
+            return XarrayDataCube(cube.get_array() + offset)
     """,
 ])
 def test_udp_udf_apply_neirghborhood_with_parameter(api100, user_defined_process_registry, set_parameters, udf_code):
@@ -2216,3 +2242,470 @@ class TestVectorCubeRunUdf:
 
     # TODO test that df index is timestamp
     # TODO: test that creates new series/dataframe with band stats
+
+
+class TestLoadResult:
+    @pytest.fixture
+    def zk_client(self) -> KazooClientMock:
+        zk_client = KazooClientMock()
+        with mock.patch(
+            "openeogeotrellis.job_registry.KazooClient", return_value=zk_client
+        ):
+            yield zk_client
+
+    @pytest.fixture
+    def zk_job_registry(self, zk_client) -> ZkJobRegistry:
+        return ZkJobRegistry(zk_client=zk_client)
+
+    @pytest.fixture
+    def urllib_mock(self) -> UrllibMocker:
+        with UrllibMocker().patch() as mocker:
+            yield mocker
+
+    def _setup_existing_job(
+        self,
+        *,
+        job_id: str,
+        api: ApiTester,
+        batch_job_output_root: Path,
+        zk_job_registry: ZkJobRegistry,
+        user_id=TEST_USER,
+    ) -> Path:
+        """
+        Set up an exiting job, with a geopyspark-driver-style job result folder,
+        (based on result files from a given job id from the test data folder),
+        and metadata in job registry.
+        """
+        source_result_dir = api.data_path(f"binary/jobs/{job_id}")
+        result_dir = batch_job_output_root / job_id
+        _log.info(f"Copy {source_result_dir=} to {result_dir=}")
+        shutil.copytree(src=source_result_dir, dst=result_dir)
+
+        # Rewrite paths in job_metadata.json
+        job_metadata_file = result_dir / JOB_METADATA_FILENAME
+        _log.info(f"Rewriting asset paths in {job_metadata_file=}")
+        job_metadata_file.write_text(
+            job_metadata_file.read_text(encoding="utf-8").replace(
+                "/data/projects/OpenEO", str(batch_job_output_root)
+            )
+        )
+
+        # Register metadata in job registry too
+        zk_job_registry.register(
+            job_id=job_id,
+            user_id=user_id,
+            api_version="1.1.0",
+            specification=load_json(result_dir / "process_graph.json"),
+        )
+        job_metadata = load_json(job_metadata_file)
+        zk_job_registry.patch(
+            job_id=job_id,
+            user_id=user_id,
+            **{k: job_metadata[k] for k in ["bbox", "epsg"]},
+        )
+        zk_job_registry.set_status(
+            job_id=job_id, user_id=user_id, status=JOB_STATUS.FINISHED
+        )
+
+        return result_dir
+
+    def test_load_result_job_id_basic(
+        self, api110, zk_client, zk_job_registry, batch_job_output_root
+    ):
+        job_id = "j-ec5d3e778ba5423d8d88a50b08cb9f63"
+
+        self._setup_existing_job(
+            job_id=job_id,
+            api=api110,
+            batch_job_output_root=batch_job_output_root,
+            zk_job_registry=zk_job_registry,
+        )
+
+        process_graph = {
+            "lc": {
+                "process_id": "load_result",
+                "arguments": {"id": job_id},
+            },
+            "save": {
+                "process_id": "save_result",
+                "arguments": {"data": {"from_node": "lc"}, "format": "json"},
+                "result": True,
+            },
+        }
+        response = api110.check_result(process_graph)
+        result = response.assert_status_code(200).json
+
+        assert result == DictSubSet(
+            {
+                "dims": ["t", "bands", "x", "y"],
+                "attrs": DictSubSet(
+                    {
+                        "shape": [3, 2, 73, 92],
+                        "crs": RegexMatcher(".*utm.*zone=31"),
+                    }
+                ),
+                "coords": DictSubSet(
+                    {
+                        "t": {
+                            "dims": ["t"],
+                            "attrs": {"dtype": "datetime64[ns]", "shape": [3]},
+                            "data": [
+                                "2022-09-07 00:00:00",
+                                "2022-09-12 00:00:00",
+                                "2022-09-19 00:00:00",
+                            ],
+                        },
+                        "bands": {
+                            "attrs": {"dtype": "<U3", "shape": [2]},
+                            "data": ["B02", "B03"],
+                            "dims": ["bands"],
+                        },
+                        "x": {
+                            "dims": ["x"],
+                            "attrs": {"dtype": "float64", "shape": [73]},
+                            "data": ListSubSet(
+                                [644765.0, 644775.0, 644785.0, 645485.0]
+                            ),
+                        },
+                        "y": {
+                            "dims": ["y"],
+                            "attrs": {"dtype": "float64", "shape": [92]},
+                            "data": ListSubSet(
+                                [5675445.0, 5675455.0, 5675465.0, 5676355.0]
+                            ),
+                        },
+                    }
+                ),
+            }
+        )
+        data = np.array(result["data"])
+        assert data.shape == (3, 2, 73, 92)
+        assert_equal(
+            data[:1, :2, :3, :4],
+            [
+                [
+                    [[584, 587, 592, 579], [580, 604, 604, 560], [610, 592, 611, 592]],
+                    [[588, 593, 572, 560], [565, 574, 580, 566], [576, 572, 590, 568]],
+                ]
+            ],
+        )
+
+    @pytest.mark.parametrize(
+        ["load_result_kwargs", "expected"],
+        [
+            (
+                {
+                    "spatial_extent": {
+                        "west": 5.077,
+                        "south": 51.215,
+                        "east": 5.080,
+                        "north": 51.219,
+                    }
+                },
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (3, 2, 24, 46),
+                    "ts": [
+                        "2022-09-07 00:00:00",
+                        "2022-09-12 00:00:00",
+                        "2022-09-19 00:00:00",
+                    ],
+                    "bands": ["B02", "B03"],
+                    "xs": [645045.0, 645055.0, 645275.0],
+                    "ys": [5675785.0, 5675795.0, 5676235.0],
+                },
+            ),
+            (
+                {"temporal_extent": ["2022-09-10", "2022-09-15"]},
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (1, 2, 73, 92),
+                    "ts": [
+                        "2022-09-12 00:00:00",
+                    ],
+                    "bands": ["B02", "B03"],
+                    "xs": [644765.0, 644775.0, 645485.0],
+                    "ys": [5675445.0, 5675455.0, 5676355.0],
+                },
+            ),
+            (
+                {"bands": ["B03"]},
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (3, 1, 73, 92),
+                    "ts": [
+                        "2022-09-07 00:00:00",
+                        "2022-09-12 00:00:00",
+                        "2022-09-19 00:00:00",
+                    ],
+                    "bands": ["B03"],
+                    "xs": [644765.0, 644775.0, 645485.0],
+                    "ys": [5675445.0, 5675455.0, 5676355.0],
+                },
+            ),
+        ],
+    )
+    def test_load_result_job_id_filtering(
+        self,
+        api110,
+        zk_client,
+        zk_job_registry,
+        batch_job_output_root,
+        load_result_kwargs,
+        expected,
+    ):
+        job_id = "j-ec5d3e778ba5423d8d88a50b08cb9f63"
+
+        self._setup_existing_job(
+            job_id=job_id,
+            api=api110,
+            batch_job_output_root=batch_job_output_root,
+            zk_job_registry=zk_job_registry,
+        )
+
+        process_graph = {
+            "lc": {
+                "process_id": "load_result",
+                "arguments": {"id": job_id, **load_result_kwargs},
+            },
+            "save": {
+                "process_id": "save_result",
+                "arguments": {"data": {"from_node": "lc"}, "format": "json"},
+                "result": True,
+            },
+        }
+        response = api110.check_result(process_graph)
+        result = response.assert_status_code(200).json
+
+        assert result["dims"] == expected["dims"]
+        assert result["attrs"]["shape"] == list(expected["shape"])
+        assert result["coords"]["t"]["data"] == expected["ts"]
+        assert result["coords"]["bands"]["data"] == expected["bands"]
+        assert result["coords"]["x"]["data"] == ListSubSet(expected["xs"])
+        assert result["coords"]["y"]["data"] == ListSubSet(expected["ys"])
+        data = np.array(result["data"])
+        assert data.shape == expected["shape"]
+
+    def _setup_metadata_request_mocking(
+        self,
+        job_id: str,
+        api: ApiTester,
+        results_dir: Path,
+        results_url: str,
+        urllib_mock: UrllibMocker,
+    ):
+        # Use ApiTester to easily build responses for the metadata request we have to mock
+        api.set_auth_bearer_token()
+        results_metadata = (
+            api.get(f"jobs/{job_id}/results").assert_status_code(200).json
+        )
+        urllib_mock.get(results_url, data=json.dumps(results_metadata))
+        # Mock each collection item metadata request too
+        for link in results_metadata["links"]:
+            if link["rel"] == "item":
+                path = link["href"].partition(api.url_root)[-1]
+                item_metadata = api.get(path).assert_status_code(200).json
+                # Change asset urls to local paths so the data can easily be read (without URL mocking in scala) by
+                # org.openeo.geotrellis.geotiff.PyramidFactory.from_uris()
+                for k in item_metadata["assets"]:
+                    item_metadata["assets"][k]["href"] = str(results_dir / k)
+                urllib_mock.get(link["href"], json.dumps(item_metadata))
+
+    def test_load_result_url_basic(
+        self,
+        api110,
+        zk_client,
+        zk_job_registry,
+        batch_job_output_root,
+        urllib_mock,
+    ):
+        job_id = "j-ec5d3e778ba5423d8d88a50b08cb9f63"
+        results_url = f"https://foobar.test/job/{job_id}/results"
+
+        results_dir = self._setup_existing_job(
+            job_id=job_id,
+            api=api110,
+            batch_job_output_root=batch_job_output_root,
+            zk_job_registry=zk_job_registry,
+        )
+        self._setup_metadata_request_mocking(
+            job_id=job_id,
+            api=api110,
+            results_dir=results_dir,
+            results_url=results_url,
+            urllib_mock=urllib_mock,
+        )
+
+        process_graph = {
+            "lc": {
+                "process_id": "load_result",
+                "arguments": {"id": results_url},
+            },
+            "save": {
+                "process_id": "save_result",
+                "arguments": {"data": {"from_node": "lc"}, "format": "json"},
+                "result": True,
+            },
+        }
+        response = api110.check_result(process_graph)
+        result = response.assert_status_code(200).json
+
+        assert result == DictSubSet(
+            {
+                "dims": ["t", "bands", "x", "y"],
+                "attrs": DictSubSet(
+                    {
+                        "shape": [3, 2, 73, 92],
+                        "crs": RegexMatcher(".*utm.*zone=31"),
+                    }
+                ),
+                "coords": DictSubSet(
+                    {
+                        "t": {
+                            "dims": ["t"],
+                            "attrs": {"dtype": "datetime64[ns]", "shape": [3]},
+                            "data": [
+                                "2022-09-07 00:00:00",
+                                "2022-09-12 00:00:00",
+                                "2022-09-19 00:00:00",
+                            ],
+                        },
+                        "bands": {
+                            "attrs": {"dtype": "<U3", "shape": [2]},
+                            "data": ["B02", "B03"],
+                            "dims": ["bands"],
+                        },
+                        "x": {
+                            "dims": ["x"],
+                            "attrs": {"dtype": "float64", "shape": [73]},
+                            "data": ListSubSet(
+                                [644765.0, 644775.0, 644785.0, 645485.0]
+                            ),
+                        },
+                        "y": {
+                            "dims": ["y"],
+                            "attrs": {"dtype": "float64", "shape": [92]},
+                            "data": ListSubSet(
+                                [5675445.0, 5675455.0, 5675465.0, 5676355.0]
+                            ),
+                        },
+                    }
+                ),
+            }
+        )
+        data = np.array(result["data"])
+        assert data.shape == (3, 2, 73, 92)
+        assert_equal(
+            data[:1, :2, :3, :4],
+            [
+                [
+                    [[584, 587, 592, 579], [580, 604, 604, 560], [610, 592, 611, 592]],
+                    [[588, 593, 572, 560], [565, 574, 580, 566], [576, 572, 590, 568]],
+                ]
+            ],
+        )
+
+    @pytest.mark.parametrize(
+        ["load_result_kwargs", "expected"],
+        [
+            (
+                {
+                    "spatial_extent": {
+                        "west": 5.077,
+                        "south": 51.215,
+                        "east": 5.080,
+                        "north": 51.219,
+                    }
+                },
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (3, 2, 24, 46),
+                    "ts": [
+                        "2022-09-07 00:00:00",
+                        "2022-09-12 00:00:00",
+                        "2022-09-19 00:00:00",
+                    ],
+                    "bands": ["B02", "B03"],
+                    "xs": [645045.0, 645055.0, 645275.0],
+                    "ys": [5675785.0, 5675795.0, 5676235.0],
+                },
+            ),
+            (
+                {"temporal_extent": ["2022-09-10", "2022-09-15"]},
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (1, 2, 73, 92),
+                    "ts": [
+                        "2022-09-12 00:00:00",
+                    ],
+                    "bands": ["B02", "B03"],
+                    "xs": [644765.0, 644775.0, 645485.0],
+                    "ys": [5675445.0, 5675455.0, 5676355.0],
+                },
+            ),
+            (
+                {"bands": ["B03"]},
+                {
+                    "dims": ["t", "bands", "x", "y"],
+                    "shape": (3, 1, 73, 92),
+                    "ts": [
+                        "2022-09-07 00:00:00",
+                        "2022-09-12 00:00:00",
+                        "2022-09-19 00:00:00",
+                    ],
+                    "bands": ["B03"],
+                    "xs": [644765.0, 644775.0, 645485.0],
+                    "ys": [5675445.0, 5675455.0, 5676355.0],
+                },
+            ),
+        ],
+    )
+    def test_load_result_url_filtering(
+        self,
+        api110,
+        zk_client,
+        zk_job_registry,
+        batch_job_output_root,
+        urllib_mock,
+        load_result_kwargs,
+        expected,
+    ):
+        job_id = "j-ec5d3e778ba5423d8d88a50b08cb9f63"
+        results_url = f"https://foobar.test/job/{job_id}/results"
+
+        results_dir = self._setup_existing_job(
+            job_id=job_id,
+            api=api110,
+            batch_job_output_root=batch_job_output_root,
+            zk_job_registry=zk_job_registry,
+        )
+        self._setup_metadata_request_mocking(
+            job_id=job_id,
+            api=api110,
+            results_dir=results_dir,
+            results_url=results_url,
+            urllib_mock=urllib_mock,
+        )
+
+        process_graph = {
+            "lc": {
+                "process_id": "load_result",
+                "arguments": {"id": results_url, **load_result_kwargs},
+            },
+            "save": {
+                "process_id": "save_result",
+                "arguments": {"data": {"from_node": "lc"}, "format": "json"},
+                "result": True,
+            },
+        }
+        response = api110.check_result(process_graph)
+        result = response.assert_status_code(200).json
+
+        assert result["dims"] == expected["dims"]
+        assert result["attrs"]["shape"] == list(expected["shape"])
+        assert result["coords"]["t"]["data"] == expected["ts"]
+        assert result["coords"]["bands"]["data"] == expected["bands"]
+        assert result["coords"]["x"]["data"] == ListSubSet(expected["xs"])
+        assert result["coords"]["y"]["data"] == ListSubSet(expected["ys"])
+        data = np.array(result["data"])
+        assert data.shape == expected["shape"]